requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.66.0
python-dotenv>=1.0.0
selenium>=4.15.0
jupyter>=1.0.0
matplotlib>=3.7.0
